# Added support for map type: as value for placeholder and in Load