keys of map will be used as value