@taylorchu 2017-03-17
@bgaifullin we cannot simply interpolate empty slice as `NULL`. Because eq empty slice and not eq empty slice will both be false, this result will be confusing.

Could you tell me what is the use case for using map to achieve this?

Thanks!