@bgaifullin 2017-03-17
@taylorchu thank you for review.

I have 2 tables. users and groups. and I have mapping. group.id -> group
I need to select all users by group.id, so I use the following code:

``` golang
type GroupsMapping map[int]Group

func (groups GroupsMapping) Build(d dbr.Dialect, b dbr.Buffer) error {
  // there is converting keys of groups into comma separated list
  // and if groups empty returns ("NULL")
}

q := s.Select(....).From("users").Where("group_id IN ?", groups)

```
IMO: it would be better if this functional will be in dbr core instead of my local project.

About . IN (NULL) and NOT IN (NULL). I agree that is confuses. I will remove this patch and add map support into dbr.Eq and dbr.Neq.