@bgaifullin 2017-03-19
@taylorchu I have updated PR