@bgaifullin 2017-04-12
Refactoring of Load method also gives +5% to performance.
```
benchmark           old ns/op     new ns/op     delta
BenchmarkLoad-4     6260          6595          +5.35%
```